backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "black>=25.1.0",
    "fastapi>=0.116.1",
    "fastapi-cors>=0.0.6",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.35.0",
]
